# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DATN API",
        "description": "API for DATN project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/supplies": {
            "get": {
                "tags": [
                    "Supplies"
                ],
                "summary": "Danh sách vật tư",
                "description": "Lấy danh sách tất cả vật tư với hỗ trợ lọc, tìm kiếm và phân trang",
                "operationId": "getSupplies",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Lọc theo danh mục",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Lọc theo trạng thái (active, inactive, discontinued)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "stock_status",
                        "in": "query",
                        "description": "Lọc theo tình trạng kho (low_stock, out_of_stock, in_stock)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm kiếm theo tên",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang (15 kết quả/trang)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách vật tư",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Supplies"
                ],
                "summary": "Tạo vật tư mới",
                "description": "Tạo một vật tư mới với thông tin chi tiết",
                "operationId": "storeSupply",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category",
                                    "unit",
                                    "current_stock",
                                    "min_stock_level",
                                    "unit_price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nước rửa bát"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nước rửa bát chuyên dụng"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Dụng cụ vệ sinh"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "Chai"
                                    },
                                    "current_stock": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "min_stock_level": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "max_stock_level": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50000
                                    },
                                    "supplier": {
                                        "type": "string",
                                        "example": "Công ty XYZ"
                                    },
                                    "supplier_contact": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "discontinued"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vật tư được tạo thành công"
                    },
                    "400": {
                        "description": "Lỗi xác thực"
                    }
                }
            }
        },
        "/supplies/{id}": {
            "get": {
                "tags": [
                    "Supplies"
                ],
                "summary": "Chi tiết vật tư",
                "description": "Lấy thông tin chi tiết của một vật tư",
                "operationId": "getSupply",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết vật tư",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vật tư không tìm thấy"
                    }
                }
            },
            "put": {
                "tags": [
                    "Supplies"
                ],
                "summary": "Cập nhật vật tư",
                "description": "Cập nhật thông tin vật tư",
                "operationId": "updateSupply",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "unit": {
                                        "type": "string"
                                    },
                                    "min_stock_level": {
                                        "type": "integer"
                                    },
                                    "max_stock_level": {
                                        "type": "integer"
                                    },
                                    "unit_price": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "discontinued"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Supplies"
                ],
                "summary": "Xóa vật tư",
                "description": "Xóa một vật tư",
                "operationId": "deleteSupply",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "400": {
                        "description": "Không thể xóa vật tư này"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "name": "Token based based security",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}